Add files via upload
Updated the data pull to include additional columns. Aggregates data for each period by game and season so that we can explore if necessary. For the stats requiring a summation by game, these are first added up before taking a season average across all games when including in the season dataset.